+++ taskManager/Helpers/Protocols/ICellViewModel.swift
@@ -13,12 +13,6 @@ protocol ICellViewAnyModel {
 	/// Настройка ячейки.
 	/// - Parameter cell: Ячейка таблицы.
 	func setupAny(cell: UIView)
-
-	/// Настройка accessibility идентификаторов
-	/// - Parameters:
-	///   - cell: Ячейка таблицы
-	///   - indexPath: Индекс ячейки
-	func setupAnyAccessibilityIds(for cell: UIView, at indexPath: IndexPath)
 }
 
 protocol ICellViewModel: ICellViewAnyModel {
@@ -27,12 +21,6 @@ protocol ICellViewModel: ICellViewAnyModel {
 	/// Настройка ячейки.
 	/// - Parameter cell: Ячейка таблицы.
 	func setup(cell: CellType)
-
-	/// Настройка accessibility идентификаторов
-	/// - Parameters:
-	///   - cell: Ячейка таблицы
-	///   - indexPath: Индекс ячейки
-	func setupAccessibilityIds(for cell: CellType, at indexPath: IndexPath)
 }
 
 extension ICellViewModel {
@@ -45,14 +33,5 @@ extension ICellViewModel {
 		}
 
 		setup(cell: cell)
-	}
-
-	func setupAnyAccessibilityIds(for cell: UIView, at indexPath: IndexPath) {
-		guard let cell = cell as? CellType else {
-			assertionFailure("Cann`t setup accessibilityIds")
-			return
-		}
-
-		setupAccessibilityIds(for: cell, at: indexPath)
 	}
 }
