+++ taskManager/Scenes/TaskList/View/Cell/ImportantTask/ImportantTaskTableViewCell.swift
@@ -5,9 +5,8 @@
 //  Created by Давид Тоноян  on 12.02.2023.
 //
 
-import UIKit
-import TaskManagerPackage
 import PinLayout
+import UIKit
 
 /// Класс ячейки для важных задач.
 final class ImportantTaskTableViewCell: UITableViewCell {
@@ -84,58 +83,22 @@ final class ImportantTaskTableViewCell: UITableViewCell {
 	}
 }
 
-/// Модель для конфигурации ячейки.
-struct ImportantTaskCellModel {
-	let taskStatus: TaskStatus
-	let taskPriority: TaskPriority
-	let taskName: String
-	let taskImportanceImage = UIImage(systemName: "exclamationmark.octagon")
-	let taskDeadline: String
-	let taskIsOverdue: Bool
-}
-
-extension ImportantTaskCellModel: ICellViewModel {
-	func setup(cell: ImportantTaskTableViewCell) {
-		cell.accessoryType = taskStatus == .completed ? .checkmark : .none
-
-		cell.taskNameLabel.text = taskName
-		cell.taskNameLabel.textColor = Theme.black
-
-		cell.taskImportanceView.image = taskImportanceImage
-		cell.taskImportanceView.tintColor = getImportanceColor(taskPriority: taskPriority)
-
-		cell.taskDeadlineLabel.text = "\(taskDeadline)"
-		cell.taskDeadlineLabel.textColor = Theme.red
-
-		cell.backgroundColor = Theme.gray
-	}
-
-	func setupAccessibilityIds(for cell: ImportantTaskTableViewCell, at indexPath: IndexPath) {
+extension ImportantTaskTableViewCell: ICellAccessibility {
+	func setupAccessibilityIds(at indexPath: IndexPath) {
 		let indexCellMark = "\(indexPath.section)_\(indexPath.row)"
 		let typeCellMark = TaskListSceneAccessibilityId.importantCell
-		cell.accessibilityIdentifier = "\(typeCellMark)_\(indexCellMark)"
+		accessibilityIdentifier = "\(typeCellMark)_\(indexCellMark)"
 
-		cell.taskNameLabel.accessibilityIdentifier =
-		"\(typeCellMark)_\(TaskListSceneAccessibilityId.taskNameLabel)_\(indexCellMark)"
+		taskNameLabel.accessibilityIdentifier =
+			"\(typeCellMark)_\(TaskListSceneAccessibilityId.taskNameLabel)_\(indexCellMark)"
 
-		cell.taskDeadlineLabel.accessibilityIdentifier =
-		"\(typeCellMark)_\(TaskListSceneAccessibilityId.taskDeadlineLabel)_\(indexCellMark)"
+		taskDeadlineLabel.accessibilityIdentifier =
+			"\(typeCellMark)_\(TaskListSceneAccessibilityId.taskDeadlineLabel)_\(indexCellMark)"
 
-		cell.taskImportanceView.accessibilityIdentifier =
-		"\(typeCellMark)_\(TaskListSceneAccessibilityId.taskImportanceView)_\(indexCellMark)"
+		taskImportanceView.accessibilityIdentifier =
+			"\(typeCellMark)_\(TaskListSceneAccessibilityId.taskImportanceView)_\(indexCellMark)"
 
-		cell.accessoryView?.accessibilityIdentifier =
-		"\(typeCellMark)_\(TaskListSceneAccessibilityId.checkMark)_\(indexCellMark)"
-	}
-
-	private func getImportanceColor(taskPriority: TaskPriority) -> UIColor {
-		switch taskPriority {
-		case .low:
-			return Theme.green
-		case .medium:
-			return Theme.yellow
-		case .high:
-			return Theme.red
-		}
+		accessoryView?.accessibilityIdentifier =
+			"\(typeCellMark)_\(TaskListSceneAccessibilityId.checkMark)_\(indexCellMark)"
 	}
 }
